Interaction Versions of Residual Diagnostics Plots.

Creates interactive versions of all residual diagnostic plots given a model. Currently accepts models of type "lm", "glm", "lmerMod", "lmerModLmerTest", and "glmerMod".

resid_interact(model, plot = "resid", type = NA, bins = NA,
  smoother = FALSE, qqline = TRUE, theme = "bw",
  axis.text.size = 10, title.text.size = 12, title.opt = TRUE)

Arguments

model
Model fit using either lm, glm, lmer, lmerTest, or glmer.

plot
Specify plot to create an interactive version. (See details for options.)

type
The user may specify a type of residuals to use. Otherwise, the default residual type for each model is used. (See details in the help file for resid_panel for options.)

bins
Number of bins to use when creating a histogram of the residuals.

smoother
Indicates whether or not to include a smoother on the residual plot. Specify TRUE or FALSE. Default is set to FALSE.

qqline
Indicates whether to include a 1-1 line on the qq-plot. Specify TRUE or FALSE. Default is set to TRUE.

theme
ggplot2 theme to be used. Current options are "bw", "classic", and "grey" (or "gray"). Default is "bw".

axis.text.size
Specifies the size of the text for the axis labels of all plots in the panel.

title.text.size
Specifies the size of the text for the titles of all plots in the panel.

title.opt
Indicates whether or not to include a title on the plots in the panel. Specify TRUE or FALSE. Default is set to TRUE.

Value

The interactive residual diagnostic plot specified.

Details

Currently, only one plot can be made interactive at a time. The options are as follows.

"boxplot": A boxplot of residuals

"cookd": A plot of Cook's D values versus observation numbers

"hist": A histogram of residuals

"index": A plot of residuals versus observation numbers

"ls": A location scale plot of the residuals

"qq": A normal quantile plot of residuals

"lev": A plot of leverage values versus residuals

"resid": A plot of residuals versus predicted values

"yvp":: A plot of observed response values versus predicted values

Note: "cookd", "ls", and "lev" are not available for " lmer", "lmerTest", and "glmer" models.

Details on the creation of the plots can be found in the details section of the help file for resid_panel.

Examples

## --------------------------------------------------------------------------------
## Linear Regression Models
## --------------------------------------------------------------------------------

# Fit a linear regression model to predict the volume of a tree based on the
# girth of a tree using the R "trees" data
lm_model1 <- lm(Volume ~ Girth, data = trees)

# Create an interactive Cook's D plot
resid_interact(lm_model1, plot = "cookd")

#> Warning: Some of the interactive plots do not function with the dev version of
#>           ggplot from github. To achieve full functionallity, please install ggplot2
#>           from CRAN.

# Fit a linear model to compare the weights of plants bewteen different
# treatment groups using the R "PlantGrowth" data
lm_model2 <- lm(weight ~ group, data = PlantGrowth)

# Create an interactive location-scale plot with them classic theme
resid_interact(lm_model2, plot = "ls", theme = "classic")

#> Warning: Some of the interactive plots do not function with the dev version of
#>           ggplot from github. To achieve full functionallity, please install ggplot2
#>           from CRAN.

## --------------------------------------------------------------------------------
## Generalized Linear Regression Models
## --------------------------------------------------------------------------------

# Fit a generalized linear regression model using a Poisson family to compare
# the insect counts between different sprays from the R "InsectSprays" data
glm_model <- glm(count ~ spray, family = "poisson", data = InsectSprays)

# Create an interactive residual plot with a smoother
resid_interact(glm_model, plot = "resid", smoother = TRUE)

#> Warning: Some of the interactive plots do not function with the dev version of
#>           ggplot from github. To achieve full functionallity, please install ggplot2
#>           from CRAN.

## --------------------------------------------------------------------------------
## Linear Mixed Effects Models
## --------------------------------------------------------------------------------

# Load the lme4 package
library(lme4)

#> Loading required package: Matrix

# Fit a linear mixed effect model to compare weights of chicks between diets using
# the R "ChickWeight" data and including chick as a random effect to account for the
# multiple measurements over time
lmer_model <- lmer(weight ~ Time + Diet + Time*Diet + (1|Chick), data = ChickWeight)

# Create an interactive qq-plot without a title
resid_interact(lmer_model, plot = "hist", title.opt = FALSE)

#> Warning: By default, bins = 30 in the histogram of residuals. If necessary, specify
#>               an appropriate number of bins.

#> Warning: Some of the interactive plots do not function with the dev version of
#>           ggplot from github. To achieve full functionallity, please install ggplot2
#>           from CRAN.

## --------------------------------------------------------------------------------
## Generalized Linear Mixed Effects Models
## --------------------------------------------------------------------------------

# Generate Poisson data
example_data <- data.frame(y = rpois(54, 3),
                           trt = rep(c("A", "B"), each = 27),
                           subject = rep(1:18, each = 3))

# Fit a generalized linear mixed effects model with a Poisson family to compare
# the response between the treatments with a random effect for subject to
# account for the dependence within a subject
glmer_model <- glmer(y ~ trt + (1|subject), family = "poisson", data = example_data)

# Create an interactive residual plot with the Pearson residuals
resid_interact(glmer_model, plot = "resid", type = "pearson")

#> Warning: Some of the interactive plots do not function with the dev version of
#>           ggplot from github. To achieve full functionallity, please install ggplot2
#>           from CRAN.